middleware/requirements.txt
feedparser==5.2.1
apscheduler==3.6.3
flask
flask-sqlalchemy==2.4.1
sqlalchemy==1.3.11
python-dateutil==2.8.1
requests==2.22.0
Flask-Cors==3.0.8
boto3==1.10.45